backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
openai>=1.30.0
supabase>=2.5.0
httpx>=0.27.0
pydantic-settings>=2.2.0
